+++ js/DOMelements.js
@@ -1,0 +1,9 @@
+const $arenas = document.querySelector('.arenas')
+const $formFigth = document.querySelector('.control')
+const $chat = document.querySelector('.chat')
+
+export {
+  $arenas,
+  $formFigth,
+  $chat
+}
