+++ js/randomGenerator.js
@@ -1,0 +1,2 @@
+const randomChanger = (num) =>  Math.ceil(Math.random() * num)
+export {randomChanger}
